Move RenderingResourcesSystem to Rendering namespace
